+++ code/src/cljs/sixsq/nuvla/ui/edge_detail/views.cljs
@@ -11,6 +11,8 @@
     [sixsq.nuvla.ui.deployment.views :as deployment-views]
     [sixsq.nuvla.ui.edge-detail.events :as events]
     [sixsq.nuvla.ui.edge-detail.subs :as subs]
+    [sixsq.nuvla.ui.edge.events :as edge-events]
+    [sixsq.nuvla.ui.edge.subs :as edge-subs]
     [sixsq.nuvla.ui.edge.utils :as utils]
     [sixsq.nuvla.ui.i18n.subs :as i18n-subs]
     [sixsq.nuvla.ui.job.subs :as job-subs]
@@ -93,6 +95,20 @@
         :selection   true}])))
 
 
+
+(defn DropdownReleases
+  [opts]
+  (let [releases (subscribe [::edge-subs/nuvlabox-releases-options])]
+    (fn [opts]
+      (when (empty? @releases)
+        (dispatch [::edge-events/get-nuvlabox-releases]))
+      [ui/Dropdown
+       (merge {:selection true
+               :loading   (empty? @releases)
+               :options   @releases}
+              opts)])))
+
+
 (defn AddRevokeSSHButton
   [{:keys [id] :as resource} operation show? title icon button-text]
   (let [tr            (subscribe [::i18n-subs/tr])
@@ -112,7 +128,7 @@
         on-error-fn   close-fn
         on-click-fn   #(do
                          (reset! loading? true)
-                         (dispatch [::events/add-revoke-ssh-key id operation @form-data
+                         (dispatch [::events/operation id operation @form-data
                                     on-success-fn on-error-fn]))]
     (fn [resource operation show? title icon button-text]
       [ui/Modal
@@ -157,6 +173,75 @@
           :loading  (true? @loading?)
           :on-click (if @key-data close-fn on-click-fn)}]]])))
 
+(defn is-old-version?
+  [nb-version]
+  (or
+    (str/blank? nb-version)
+    (let [p (->> (str/split nb-version #"\.")
+                 (map js/parseInt))]
+      (or (< (first p) 1)
+          (and (= (first p) 1)
+               (< (second p) 14))))))
+
+(defn UpdateButton
+  [{:keys [id] :as resource} operation show? title icon button-text]
+  (let [tr            (subscribe [::i18n-subs/tr])
+        status        (subscribe [::subs/nuvlabox-status])
+        close-fn      #(reset! show? false)
+        form-data     (r/atom nil)
+        on-change-fn  #(swap! form-data assoc :nuvlabox-release %)
+        on-success-fn close-fn
+        on-error-fn   close-fn
+        on-click-fn   #(dispatch [::events/operation id operation @form-data
+                                  on-success-fn on-error-fn])]
+    (fn [{:keys [id] :as resource} operation show? title icon button-text]
+      (when (not= (:parent @status))
+        (dispatch [::events/get-nuvlabox id]))
+      (let [correct-nb? (= (:parent @status) id)
+            nb-version  (get @status :nuvlabox-engine-version "")]
+        (when-not correct-nb?
+          ;; needed to make modal work in cimi detail page
+          (dispatch [::events/get-nuvlabox id]))
+        [ui/Modal
+         {:open       @show?
+          :close-icon true
+          :on-close   close-fn
+          :trigger    (r/as-element
+                        [ui/MenuItem {:on-click #(reset! show? true)}
+                         [ui/Icon {:name icon}]
+                         title])}
+         [ui/ModalHeader [:div title]]
+         [ui/ModalContent
+          (when correct-nb?
+            [:<>
+             (when (is-old-version? nb-version)
+               [ui/Message
+                {:warning true
+                 :header  "NuvlaBox update warning"
+                 :content (r/as-element
+                            [:span (str "Your NuvlaBox version is older than v1.14. "
+                                        "The update operation might not work as expected. ")
+                             [:a {:href   (str "https://docs.nuvla.io/nuvlabox/"
+                                               "nuvlabox-engine/quickstart.html#from-nuvla")
+                                  :target "_blank"}
+                              "See more"]])}])
+             [ui/Segment
+              [:b "Current Engine Version: "]
+              [:i nb-version]]])
+          [ui/Segment
+           [:b "Update to: "]
+           [DropdownReleases {:placeholder "select a version"
+                              :on-change   (ui-callback/value #(on-change-fn %))}]]]
+         [ui/ModalActions
+          [uix/Button
+           {:text     (@tr [:cancel])
+            :on-click close-fn}]
+          [uix/Button
+           {:text     button-text
+            :disabled (str/blank? (:nuvlabox-release @form-data))
+            :primary  true
+            :on-click on-click-fn}]]]))))
+
 
 (defmethod cimi-detail-views/other-button ["nuvlabox" "add-ssh-key"]
   [resource operation]
@@ -175,6 +260,15 @@
       [AddRevokeSSHButton resource operation show? "Revoke ssh key" "minus" "revoke"])))
 
 
+(defmethod cimi-detail-views/other-button ["nuvlabox" "update-nuvlabox"]
+  [resource operation]
+  (let [tr    (subscribe [::i18n-subs/tr])
+        show? (r/atom false)]
+    (fn [resource operation]
+      ^{:key (str "update-nuvlabox" @show?)}
+      [UpdateButton resource operation show? "Update NuvlaBox" "download" (@tr [:update])])))
+
+
 (defn MenuBar [uuid]
   (let [can-decommission? (subscribe [::subs/can-decommission?])
         can-delete?       (subscribe [::subs/can-delete?])
@@ -188,7 +282,8 @@
         (when @can-delete?
           [DeleteButton @nuvlabox])
 
-        [cimi-detail-views/format-operations @nuvlabox #{"decommission" "commission" "check-api"}]
+        [cimi-detail-views/format-operations @nuvlabox #{"edit" "delete" "activate" "decommission"
+                                                         "commission" "check-api"}]
 
         [main-components/RefreshMenu
          {:action-id  refresh-action-id