+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+  - Fix/update broken external links
   - Edge detail - Modal update warning message for old NBE
 
 ## [2.13.2] - 2021-02-10